refactor(web): name the toggle helper module for its export
Rename lib/set.ts to lib/toggle-in-set.ts. "set" read like a Set
data-structure library; the module holds one immutable-toggle helper
over the native Set, so name the file for it.
